% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cordex_local.R
\name{get_cordex_local}
\alias{get_cordex_local}
\title{Get CORDEX driver data}
\usage{
get_cordex_local(lon, lat, scenario = c("historical", "rcp26", "rcp45",
  "rcp85"), model = "NAM-44_CCCma-CanESM2")
}
\description{
Returns full time series of CORDEX data based on requested model and lat/lon.
}
