% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_storage.R
\name{get_local_storage}
\alias{get_local_storage}
\title{Return local storage path}
\usage{
get_local_storage()
}
\description{
NLDAS data can be stored locally. The
package uses an internally-cached path, which can 
be accessed and defined using get_* and set_local_storage.
}

