% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_primary_forcing_local_multitool.R
\name{get_primary_forcing_local_multitool}
\alias{get_primary_forcing_local_multitool}
\title{Fast multi-lat/lon driver lookup}
\usage{
get_primary_forcing_local_multitool(lons, lats, ids)
}
\arguments{
\item{lons}{Vector of longitudes}

\item{lats}{Vector of latitudes}

\item{ids}{Site ids for mapping of data to lats/lons}
}
