% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_primary_forcing_rods.R
\name{nldas2_primary_forcing_rods}
\alias{nldas2_primary_forcing_rods}
\title{Get NLDAS primary forcing data}
\usage{
nldas2_primary_forcing_rods(lon, lat, start = as.POSIXct("1979-01-02", tz =
  "UTC"), stop = as.POSIXct("2017-01-01", tz = "UTC"))
}
\arguments{
\item{lon}{Longitude of data location request}

\item{lat}{Latitude of data location request}

\item{start}{Start date as POSIXct}

\item{stop}{End date as POSIXct}
}
\description{
Interfaces with NASA Data Rods (Time series) services managed by Goddard Earth Sciences Data and
Information Services Center (GES DISC). More info about the interface available
\href{https://disc.sci.gsfc.nasa.gov/hydrology/data-rods-time-series-data}{here}.
}

