% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_storage.R
\name{set_local_storage}
\alias{set_local_storage}
\title{Set local storage path}
\usage{
set_local_storage(path)
}
\arguments{
\item{path}{Path to NLDAS cache directory}
}
\description{
NLDAS data can be stored locally. The
package uses an internally-cached path, which can 
be accessed and defined using get_* and set_local_storage.
}

